+++ CHANGES.md
@@ -7,6 +7,7 @@
     its completed tasks.
 -   Check status view now also correctly updates the status
 -   Added Chloride to fieldwork options
+-   Bugfix: write errors to bro_errors instead of the log on check_status
 
 
 ## 1.57 (2025-11-12)